Move `fix_paginate` into nyaa.extensions
